Renamed utility files, removed unnecessary test
